LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-works-on-frame-ancestors.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://127.0.0.1:9443/content-security-policy/reporting-api/support/non-embeddable-frame.html?id=246bfa1d-fd81-4e9e-b32e-6a6c4ce151a8 because it does not appear in the frame-ancestors directive of the Content Security Policy.


FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
FAIL No report to parent. assert_equals: CSP report sent, but not expecting one. expected 0 but got 1
